Update h2d.py
Used a formatter: yapf
Also, fixed a bug in the links section. If does not had an 'href', it had problems.

Also, changed the table creation, so checks the max number of columns, when you have merged elements.

Added a fix in the color elements in case it has some trash (I got an html with \n all over).